rename init to initialize in qobj
