+++ qiskit/transpiler/passes/mapping/barrier_before_final_measurements.py
@@ -51,12 +51,10 @@ class BarrierBeforeFinalMeasurements(TransformationPass):
         for creg in dag.cregs.values():
             barrier_layer.add_creg(creg)
 
-        final_qubits = set(final_op.qargs[0]
-
-                           for final_op in final_ops)
+        final_qubits = set(final_op.qargs[0] for final_op in final_ops)
 
         new_barrier_node = barrier_layer.apply_operation_back(
-            Barrier(len(final_qubits)), final_qubits, [])
+            Barrier(len(final_qubits)), list(final_qubits), [])
 
         # Preserve order of final ops collected earlier from the original DAG.
         ordered_final_nodes = [node for node in dag.nodes_in_topological_order()
